When updating packages, use the greatest version
When running npm-check-updates, find the greatest versions available
instead of the latest stable versions. This allows updating some of the
beta packages in use. Since packages are not committed until release
notes are reviewed, it is better to be aware that updates exist and to
skip them than to not even see them.